Move extra legacy code to deprecating package
