+++ ClickHaskell/ClickHaskell.cabal
@@ -40,8 +40,6 @@ Library
   Exposed-Modules:
     ClickHaskell
     ClickHaskell.DbTypes
-    ClickHaskell.Reading
-    ClickHaskell.Writing
     ClickHaskell.NativeProtocol.ClientPackets
     ClickHaskell.NativeProtocol.Serialization
     ClickHaskell.NativeProtocol.ServerPackets
